# swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "Learning Node with Express",
      "description": "Building a restful Api with Swagger UI integration for docs",
      "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "tags": [
      {
        "name": "Courses",
        "description": "API for Courses in the system"
      }
    ],
    "schemes": [
      "http"
    ],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "paths": {
      "/courses": {
        "post": {
          "tags": [
            "Courses"
          ],
          "description": "Add a course to the system",
          "parameters": [
            {
              "name": "course",
              "in": "body",
              "description": "Add a course",
              "schema": {
                "$ref": "#/definitions/Course"
              }
            }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "New course is created",
              "schema": {
                "$ref": "#/definitions/Course"
              }
            }
          }
        },
        "get": {
          "tags": [
            "Courses"
          ],
          "summary": "Get all courses in system",
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/Courses"
              }
            }
          }
        }
      },
      "/courses/{id}": {
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of course that we want to find",
            "type": "string"
          }
        ],
        "get": {
          "tags": [
            "Courses"
          ],
          "summary": "Get course with given ID",
          "responses": {
            "200": {
              "description": "Course is found",
              "schema": {
                "$ref": "#/definitions/Course"
              }
            }
          }
        },
        "delete": {
          "summary": "Delete course with given ID",
          "tags": [
            "Courses"
          ],
          "responses": {
            "200": {
              "description": "Course is deleted",
              "schema": {
                "$ref": "#/definitions/updateCourse"
              }
            }
          }
        },
        "put": {
          "summary": "Update course with give ID",
          "tags": [
            "Courses"
          ],
          "parameters": [
            {
              "name": "course",
              "in": "body",
              "description": "course with new values of properties",
              "schema": {
                "$ref": "#/definitions/Course"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Course is updated",
              "schema": {
                "$ref": "#/definitions/Course"
              }
            }
          }
        }
      }
    },
    "definitions": {
      "Course": {
        "required": [
          "name"
        ],
        "properties": {
             "name": {
            "type": "string",
            "uniqueItems": true
          }
        }
      },
      "updateCourse": {
        "required": [
          "_id"
        ]
    },
      "Courses": {
        "type": "array",
        "$ref": "#/definitions/Course"
      }
    }
  }